[README] Updates readme display dependencies for my setup
Finding it a bit annoying to have to figure out which dependencies are
missing one-by-one, when doing a new setup. Instead, I think it will be
much easier if I just list off the packages I use in the readme, so that
I can just install them all in one go before starting.