[TMUX] Fixes configuration mistake
Sets the leader binding back to the default, but keeps the comments
there. I'm still thinking I'd prefer <C - a> as my leader for tmux, as
it's a little easier to hit. Although, <C - a> is also used for moving
the cursor to the start of the line... something for which I have not
found another alternative withing tmux.

Going to have to revert for now while I figure that one out.

Also, it looks like the vertical split hotkey got removed by mistake
somehow... adding it back!