[WEECHAT] Updates after using the client for a while
Played around with some of the settings after using the client for a few
days... still not really sure what kind of configuration I'm looking for
but I think this is still a good starting point.

Will probably be able to think of a better commit message as I figure
out what I want... heh.