[TMUX] Changes tmux leader from <C-b> to <C-a>
This is typically recommended in most tmux tutorials I've seen. I
initially ignored it, as I dismissed the hotkey as being a legacy habit,
coming from those who used to use screen.

However, after giving it a try, it is *much* easier to use <C-a>.
Especially with caps-lock bound to ctrl/esc (code keyboard w/ xcape).

Going to leave this configured this way!