[XPROFILE] Adds xcape and xbindkeys to xprofile
This show only start xcape and xbindkeys once. While this doesn't start
them in cases where x isn't running, I think this is good enough of a
compromise for now.

Could make this better by finding a configuration that will start with a
given tty. Not really sure if that is even a thing, or something you'd
even want to do in the first place.

Works for now!