[VIM] Allow moving buffers into "background"
Not really sure if the summary is the right terminology.

When I first started using vim, I wasn't sure how I felt about this
setting. Now, I feel that hiding buffers is very convenient, now that I
better understand (and use) the jump list.

In addition to the jump list, you can also use FZF to see and switch
between your active buffers. Not having `set hidden` sort-of gets in the
way of that feature.

Overall, I'm deciding that this is what I prefer.