[VIMRC] Updates Gruvbox theme settings
Moved some of the variable definitions to the plugin inclusion. I've
been organizing this file by keeping all of the variable definitions
close to where the plugin is included. That way it's easy to follow the
repo link to find+modify the settings, if need be. It just makes sense
to have those be next to each other.

Also, tweaks a few settings about the Gruvbox theme -- some changes I've
been wanting for a little bit, which make things look a little more to
my tastes. Italics and removing the bg on the sign column, which I found
to be ugly/distracting.