[VIM] Hotkey updates
I did not find much use for the H and L hotkeys. I've found that I like
the vim defaults, but instead prefer "larger jumps" through the CTRL
modifier.

I've already been using C-jk to jump further steps. This will extend
that pattern to the entire C-hjkl mappings. Now you can jump "further"
with C-hl, if that makes sense.

Also, adds a QOL binding for removing the Q binding, which I do not use
and found to be annoying.